+++ protostuff-runtime/src/main/java/io/protostuff/runtime/RuntimeEnv.java
@@ -38,10 +38,10 @@ public final class RuntimeEnv
     public static final boolean AUTO_LOAD_POLYMORPHIC_CLASSES;
 
     /**
-     * Disabled by default. Writes a sentinel value (uint32) in place of null values. Works only on the binary formats
-     * (protostuff/graph/protobuf).
+     * Disabled by default. Writes a sentinel value (uint32) in place of null values in collections/arrays.
+     * Works only on the binary formats (protostuff/graph/protobuf).
      */
-    public static final boolean ALLOW_NULL_ARRAY_ELEMENT;
+    public static final boolean PRESERVE_NULL_ELEMENTS;
 
     /**
      * Disabled by default. For pojos that are not declared final, they could still be morphed to their respective
@@ -185,8 +185,8 @@ public final class RuntimeEnv
         AUTO_LOAD_POLYMORPHIC_CLASSES = Boolean.parseBoolean(props.getProperty(
                 "protostuff.runtime.auto_load_polymorphic_classes", "true"));
 
-        ALLOW_NULL_ARRAY_ELEMENT = Boolean.parseBoolean(props.getProperty(
-                "protostuff.runtime.allow_null_array_element", "false"));
+        PRESERVE_NULL_ELEMENTS = Boolean.parseBoolean(props.getProperty(
+                "protostuff.runtime.preserve_null_elements", "false"));
 
         MORPH_NON_FINAL_POJOS = Boolean.parseBoolean(props.getProperty(
                 "protostuff.runtime.morph_non_final_pojos", "false"));